requirements.txt
google-generativeai
numpy<2.0.0
pyarrow>=14.0.0,<15.0.0
pandas>=2.1.0,<2.2.0
requests
pillow
streamlit
"fuzzywuzzy[speedup]"